2025.04.09

+++ src/post/service.py
@@ -3,10 +3,17 @@ from post.model.model import PostModel
 from post.schemas.request import CreatePostSchema, UpdatePostSchema, PaginatePostSchema
 from post.schemas.response import PostSchema, PaginatedPostSchema
 from post.repository import PostRepository
+from post.image.image_service import PostImageService
+import os
+import logging
+from common.const.path_consts import POST_IMAGE_PATH  # 이미지 저장 경로도 import 필요
+
+logger = logging.getLogger(__name__)
 
 class PostService:
     def __init__(self):
         self.post_repo = PostRepository()  # repo도 동일 세션을 context에서 사용 트랜잭션을 위해서
+        self.post_image_service = PostImageService()
 
     async def generate_dummy_posts(self):
         await self.post_repo.generate_posts()
@@ -44,15 +51,34 @@ class PostService:
         # Pydantic 모델로 만들어서 리턴(검증, json으로 변환 해줌)
         return PostSchema.model_validate(saved)
 
-    async def update_post(self, post_id: int, data: UpdatePostSchema):
-        # Pydantic 모델(CreateToDoReqeust)로 받은 것을 SQLAlchemy ORM 객체로 변환 -> 이렇게 해야지 DB에 들어감
-        updated = PostModel.update(request=data)
+    async def update_post_content(self, user_id: int, post_id: int, data: UpdatePostSchema):
+        post = await self.post_repo.get_post_by_id(post_id)
+        if not post:
+            raise HTTPException(status_code=404, detail="존재하지 않는 게시글입니다.")
+        if post.author_id != user_id:
+            raise HTTPException(status_code=403, detail="수정 권한이 없습니다.")
 
-        # SQLAlchemy ORM 객체로 변환된 것을 가지고 DB에 저장
-        post = await self.post_repo.update_post(post_id, updated)
-
-        # Pydantic 모델로 만들어서 리턴(검증, json으로 변환 해줌)
-        return PostSchema.model_validate(post)
-
+        await self.post_repo.update_post(post, data)
+    
     async def delete_post(self, post_id: int):
+        # 1. 게시글 조회
+        post = await self.post_repo.get_post_by_id(post_id)
+        if not post:
+            raise HTTPException(status_code=404, detail="게시글이 존재하지 않습니다.")
+
+        # 2. 이미지 경로 수집
+        deleted_image_paths = []
+        for image in post.images:
+            path = os.path.join(POST_IMAGE_PATH, image.path)
+            deleted_image_paths.append(path)
+
+        # 3. DB 삭제
         await self.post_repo.delete_post(post_id)
+
+        # 4. 디스크 이미지 삭제 (DB 성공 이후)
+        for path in deleted_image_paths:
+            if os.path.exists(path):
+                try:
+                    os.remove(path)
+                except Exception as e:
+                    logger.warning(f"⚠️ 이미지 파일 삭제 실패: {path} - {e}")
